+++ src/main/java/Model/AbstractId.java
@@ -9,13 +9,13 @@ import jakarta.persistence.MappedSuperclass;
 public class AbstractId {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
